part2/requirements.txt
python-pptx==0.6.23
fastapi==0.99.0
mangum==0.17.0
openai==1.8.0
reportlab==4.0.9
requests
python-multipart==0.0.9
uvicorn==0.26.0
aiofiles==23.2.1
boto3==1.34.130
pydantic==1.10.15
pillow==10.2.0
loguru==0.7.2



